+++ src/main/java/club/iananderson/seasonhud/client/minimaps/XaeroMinimap.java
@@ -87,8 +87,8 @@ public class XaeroMinimap {
             float left = 6 + iconDim;
             float right = (int) (mapSize + 2 + padding - stringWidth);
 
-            float stringX = scaledX + (align == 0 ? center : (align == 1 ? left : right));
-            float stringY = scaledY + (under ? mapSize + (2 * padding) : -9) + (trueCount * stringHeight * (under ? 1 : -1));
+            int stringX = (int)(scaledX + (align == 0 ? center : (align == 1 ? left : right)));
+            int stringY = (int)(scaledY + (under ? mapSize + (2 * padding) : -9) + (trueCount * stringHeight * (under ? 1 : -1)));
 
             if ((!modMain.getSettings().hideMinimapUnderScreen || mc.screen == null || mc.screen instanceof IScreenBase || mc.screen instanceof ChatScreen || mc.screen instanceof DeathScreen)
                     && (!modMain.getSettings().hideMinimapUnderF3 || !mc.options.renderDebug) && modMain.getSettings().getMinimap()) {
